+++ OpenMarket/OpenMarket/Controller/ViewController/AddItemViewController.swift
@@ -7,16 +7,6 @@
 
 import UIKit
 
-struct ItemComponents {
-    let name: String
-    let price: Int
-    let currency: String
-    let discountedPrice: Int
-    let stock: Int
-    let descriptions: String
-    let imageArray: [UIImage]
-}
-
 protocol AddItemViewControllerDelegate {
     func upDate()
 }
