+++ OpenMarket/OpenMarket/Model/ItemModel/ItemComponents.swift
@@ -1,0 +1,18 @@
+//
+//  ItemComponents.swift
+//  OpenMarket
+//
+//  Created by 두기, minseong on 2022/05/31.
+//
+
+import UIKit
+
+struct ItemComponents {
+    let name: String
+    let price: Int
+    let currency: String
+    let discountedPrice: Int
+    let stock: Int
+    let descriptions: String
+    let imageArray: [UIImage]
+}
